fix: numpy serialization, ddgs migration, full agent test coverage
- embeddings.py: coerce numpy arrays to native Python floats (fixes
  500 on GET /api/v1/content — numpy.float32 not JSON-serializable)
- recommendation_agent.py: explicit float() cast on _relevance_score
- course_planner.py: migrate from deprecated duckduckgo_search to ddgs
- pyproject.toml: remove all version pins; keep only huggingface-hub>=0.29.0
- uv.lock: regenerated (duckduckgo-search→ddgs, loose constraints)
- test_hf.py: add full test coverage for image_captioner, speech_to_text,
  embeddings, difficulty_scorer, recommendation_agent, spaced_repetition
- test_e2e.py: remove broken SQLAlchemy create_all_tables fixture import

All 170 tests pass.

+++ backend/app/hf/recommendation_agent.py
@@ -93,7 +93,7 @@ async def rank_content_for_learner(
         for i, item in enumerate(items_to_score):
             updated = dict(item)
             updated["is_ai_recommended"] = i in top_indices
-            updated["_relevance_score"] = round(scored[next(j for j, (idx, _) in enumerate(scored) if idx == i)][1], 4) if i < len(scored) else 0.0
+            updated["_relevance_score"] = float(round(scored[next(j for j, (idx, _) in enumerate(scored) if idx == i)][1], 4)) if i < len(scored) else 0.0
             result.append(updated)
 
         # Append remaining items (beyond 20) unchanged
